+++ ChangeLog.md
@@ -5,6 +5,8 @@
 - add MenuItemVM.Command
 - add MenuItemVM.CommandParameter
 - add WindowVM.ClosingEvent
+- DataTypeConverter.Convert: fix System.InvalidCastException
+- DebugType=full and DebugSymbols=true for each configuration 
 
 # 0.18
 - split KsWare.Presentation.Behavior